--- a/src/physics/grid/grid.hpp
+++ b/src/physics/grid/grid.hpp
@@ -12,8 +12,8 @@ class Grid {
 private:
   // TODO (enriavil1): add variable width and height cells by taking the biggest
   // size of width and height from objects
-  const uint32_t m_width = 50;
-  const uint32_t m_height = 50;
+  const uint32_t m_width = 20;
+  const uint32_t m_height = 20;
 
   std::unordered_map<uint32_t, GridCell> m_cells =
       std::unordered_map<uint32_t, GridCell>{};
@@ -25,35 +25,37 @@ public:
   uint32_t getHeight() { return this->m_height; }
 
   uint32_t getCellID(uint32_t pos_x, uint32_t pos_y) {
+    const auto window_width = ImGui::GetWindowWidth();
     const auto window_height = ImGui::GetWindowHeight();
     const auto window_pos = ImGui::GetWindowPos();
 
-    const auto amount_of_columns = (window_height / m_height);
+    const uint32_t amount_of_columns = (window_width / this->m_width) + 1;
+    const uint32_t amount_of_rows = (window_height / this->m_height) + 1;
 
-    return ((pos_y - window_pos.y) / this->m_height) * amount_of_columns +
-           ((pos_x - window_pos.x) / this->m_width);
+    uint32_t row = ((pos_y - window_pos.y) / this->m_height);
+    uint32_t column = ((pos_x - window_pos.x) / this->m_width);
+
+    // making sure the row and column dont leave the max range
+    row = std::min(row, amount_of_rows - 1);
+    column = std::min(column, amount_of_columns - 1);
+
+    return row * amount_of_columns + column;
   }
 
   GridCell& getCell(uint32_t id) { return this->m_cells[id]; }
 
   GridCell& getCell(uint32_t pos_x, uint32_t pos_y) {
     const uint32_t id = getCellID(pos_x, pos_y);
-
     return this->m_cells[id];
   }
 
   GridCell& getCell(PhysicsObject *obj) {
     const uint32_t id = getCellID(obj->getPosition().x, obj->getPosition().y);
-
     return this->m_cells[id];
   }
 
   void add(PhysicsObject *obj) {
     const uint32_t id = getCellID(obj->getPosition().x, obj->getPosition().y);
-
-    if (this->m_cells.find(id) == this->m_cells.end()) {
-      this->m_cells.insert(std::make_pair(id, GridCell()));
-    }
 
     this->m_cells[id].add(obj);
   }