--- a/src/physics/grid/grid.hpp
+++ b/src/physics/grid/grid.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <iostream>
+#include <unordered_map>
 #include <utility>
 #include <vector>