--- a/src/physics/grid/gridCell.hpp
+++ b/src/physics/grid/gridCell.hpp
@@ -1,0 +1,14 @@
+#pragma once
+
+#include <vector>
+
+#include "../physicsObjects/physicsObject.hpp"
+
+class GridCell {
+private:
+  std::vector<PhysicsObject *> m_objects = std::vector<PhysicsObject *>{};
+
+public:
+  std::vector<PhysicsObject *> getObjects() { return this->m_objects; }
+  void add(PhysicsObject *obj) { m_objects.push_back(obj); }
+};