Added a grid system to allow more balls without dropping frames.

Went from:
 1200 balls to drop from 120fps to 60fps
1700 balls from 60fps to 40fps

To:
3000 balls to drop from 120fps to 60fps
4800 balls to drop from 60fps to 40fps